requirements.txt
django==1.7.4
djangorestframework
